add tests

+++ 0x1E-search_algorithms/0-linear.c
@@ -18,8 +18,8 @@ int linear_search(int *array, size_t size, int value)
 
 	for (i = 0; i < (int)size; i++)
 	{
-		printf("value checked array[%u] = [%d]\n", i, array[i]);
-		if (value == array[i])
+		printf("value checked array[%u] = [%d]\n", (int)i, array[i]);
+		if (value == array[(int)i])
 			return ((int)i);
 	}
 	return (-1);
